% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{ATAC_net}
\alias{ATAC_net}
\title{Results of ConNetGNN() for scATAC-seq data from SNARE-seq dataset}
\format{
a list of three adjacency matrices.
}
\usage{
ATAC_net
}
\description{
A list to store the gene association network of scATAC-seq data.
Case data from the SNARE-seq dataset.
}
\examples{
data(ATAC_net)
}
\keyword{datasets}
