% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTMG.R
\name{BIC_LTMG}
\alias{BIC_LTMG}
\title{BIC_LTMG}
\usage{
BIC_LTMG(y, rrr, Zcut)
}
\arguments{
\item{y}{Internal parameters.}

\item{rrr}{Internal parameters.}

\item{Zcut}{Internal parameters.}
}
\description{
The internal functions of the \code{scapGNN} package.
}
\details{
BIC_LTMG
}
