% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTMG.R
\name{BIC_ZIMG}
\alias{BIC_ZIMG}
\title{BIC_ZIMG}
\usage{
BIC_ZIMG(y, rrr, Zcut)
}
\arguments{
\item{y}{Internal parameters.}

\item{rrr}{Internal parameters.}

\item{Zcut}{Internal parameters.}
}
\description{
The internal functions of the \code{scapGNN} package.
}
\details{
BIC_ZIMG
}
