% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{ConNetGNN_data}
\alias{ConNetGNN_data}
\title{The results of ConNetGNN() function}
\format{
a list.
}
\usage{
ConNetGNN_data
}
\description{
Results of ConNetGNN() function with Hv_exp as input.
}
\examples{
data(ConNetGNN_data)
}
\keyword{datasets}
