% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_scapGNN_env}
\alias{create_scapGNN_env}
\title{Create the create_scapGNN_env environment on miniconda}
\usage{
create_scapGNN_env()
}
\description{
The internal functions of the \code{scapGNN} package.
}
\details{
create_scapGNN_env
}
