% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTMG.R
\name{Fit_LTMG}
\alias{Fit_LTMG}
\title{Fitting function for Left-truncated mixed Gaussian}
\usage{
Fit_LTMG(x, n, q, k, err = 1e-10)
}
\arguments{
\item{x}{Internal parameters.}

\item{n}{Internal parameters.}

\item{q}{Internal parameters.}

\item{k}{Internal parameters.}

\item{err}{Internal parameters.}
}
\description{
The internal functions of the \code{scapGNN} package.
}
\details{
Fit_LTMG
}
