% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTMG.R
\name{Global_Zcut}
\alias{Global_Zcut}
\title{Global_Zcut}
\usage{
Global_Zcut(MAT, seed = 123)
}
\arguments{
\item{MAT}{Internal parameters.}

\item{seed}{Random number generator seed.}
}
\description{
The internal functions of the \code{scapGNN} package.
}
\details{
Global_Zcut
}
