% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{H9_0h_cpGM_data}
\alias{H9_0h_cpGM_data}
\title{Cell-activated gene modules under the 0-hour phenotype}
\format{
a list.
}
\usage{
H9_0h_cpGM_data
}
\description{
Results of cpGModule() function.
}
\examples{
data(H9_0h_cpGM_data)
}
\keyword{datasets}
