% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{Hv_exp}
\alias{Hv_exp}
\title{Single-cell gene expression profiles}
\format{
a matrix.
}
\usage{
Hv_exp
}
\description{
A log-transformed gene-cell matrix containing highly variable features.
}
\examples{
data(Hv_exp)
}
\keyword{datasets}
