% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{instPyModule}
\alias{instPyModule}
\title{Install the pyhton module through the reticulate R package}
\usage{
instPyModule(module)
}
\arguments{
\item{module}{Internal parameters.}
}
\description{
The internal functions of the \code{scapGNN} package.
}
\details{
instPyModule
}
