% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isLoaded}
\alias{isLoaded}
\title{The internal functions of the \code{scapGNN} package}
\usage{
isLoaded(name)
}
\arguments{
\item{name}{Internal parameters.}
}
\description{
Determine if the package is loaded.
}
\details{
isLoaded
}
