% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_path_data}
\alias{load_path_data}
\title{load pathway or gene set's gmt file}
\usage{
load_path_data(gmt_file_path)
}
\arguments{
\item{gmt_file_path}{Internal parameters.}
}
\value{
a list
}
\description{
The internal functions of the \code{scapGNN} package.

file format:
1. first index: pathway's name or ID.
2. second index: pathway's url or others, it dosen't matter.
3. third to all: gene symbols in pathway.
}
\details{
load_path_data
}
