% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTMG.R
\docType{class}
\name{LTMG-class}
\alias{LTMG-class}
\title{An S4 class to represent the input data for LTMG.}
\description{
An S4 class to represent the input data for LTMG.
}
\section{Slots}{

\describe{
\item{\code{InputData}}{Input data for LTMG.}

\item{\code{OrdinalMatrix}}{LTMG output data.}
}}

