% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTMG.R
\name{LTMG}
\alias{LTMG}
\title{Left-truncated mixed Gaussian}
\usage{
LTMG(VEC, Zcut_G, k = 5)
}
\arguments{
\item{VEC}{Internal parameters.}

\item{Zcut_G}{Internal parameters.}

\item{k}{Internal parameters.}
}
\description{
Functional implementation of Left-truncated mixed Gaussian. The internal functions of the \code{scapGNN} package.
}
\details{
LTMG
}
