% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{RNA_ATAC_IntNet}
\alias{RNA_ATAC_IntNet}
\title{Results of InteNet() for integrating scRNA-seq and scATAC-seq data.}
\format{
a list of three adjacency matrices.
}
\usage{
RNA_ATAC_IntNet
}
\description{
An integrated network of scRNA-seq and scATAC-seq data from SNARE-seq.
}
\examples{
data(RNA_ATAC_IntNet)
}
\keyword{datasets}
