% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{RNA_net}
\alias{RNA_net}
\title{Results of ConNetGNN() for scRNA-seq data from SNARE-seq dataset}
\format{
a list of three adjacency matrices.
}
\usage{
RNA_net
}
\description{
A list to store the gene association network of scRNA-seq data.
Case data from the SNARE-seq dataset.
}
\examples{
data(RNA_net)
}
\keyword{datasets}
