% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTMG.R
\name{RunLTMG}
\alias{RunLTMG}
\alias{.RunLTMG}
\alias{RunLTMG,LTMG-method}
\title{Run Left-truncated mixed Gaussian}
\usage{
.RunLTMG(object, Gene_use = NULL, k = 5, verbose, seed = 123)

RunLTMG(object, Gene_use = NULL, k = 5, verbose, seed = 123)

\S4method{RunLTMG}{LTMG}(object, Gene_use = NULL, k = 5, verbose, seed = 123)
}
\arguments{
\item{object}{A LTMG object}

\item{Gene_use}{using X numebr of top variant gene. input a number, recommend 2000.}

\item{k}{Constant, defaults \code{5}.}

\item{verbose}{Gives information about each step.}

\item{seed}{Random number generator seed.}
}
\value{
A list contains raw input data and LTMG results.
}
\description{
Functional implementation of Left-truncated mixed Gaussian. The internal functions of the \code{scapGNN} package.
}
\details{
RunLTMG

For more information, please refer to: \code{DOI: 10.1093/nar/gkz655} and \code{https://github.com/zy26/LTMGSCA}.
}
