% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{scPathway_data}
\alias{scPathway_data}
\title{Single cell pathway activity matrix}
\format{
a matrix.
}
\usage{
scPathway_data
}
\description{
Results of scPathway() function.
}
\examples{
data(scPathway_data)
}
\keyword{datasets}
